refactor: drop yaml for now
