+++ package.json
@@ -41,7 +41,6 @@
 		"js-cookie": "^3.0.5",
 		"lucide-react": "^1.8.0",
 		"monaco-editor": "^0.55.1",
-		"monaco-yaml": "^5.4.1",
 		"next-themes": "^0.4.6",
 		"openapi-fetch": "^0.17.0",
 		"papaparse": "^5.5.3",
@@ -57,7 +56,6 @@
 		"sonner": "^2.0.7",
 		"tailwind-merge": "^3.5.0",
 		"tw-animate-css": "^1.4.0",
-		"yaml": "^2.8.3",
 		"zod": "^4.3.6"
 	},
 	"devDependencies": {
