+++ src/modules/monaco/util/setup.ts
@@ -1,11 +1,13 @@
 import { loader } from "@monaco-editor/react";
 import * as monaco from "monaco-editor";
 import editorWorker from "monaco-editor/esm/vs/editor/editor.worker?worker";
-import yamlWorker from "monaco-yaml/yaml.worker?worker";
+import jsonWorker from "monaco-editor/esm/vs/language/json/json.worker?worker";
 
 self.MonacoEnvironment = {
 	getWorker(_, label) {
-		if (label === "yaml") return new yamlWorker();
+		if (label === "json") {
+			return new jsonWorker();
+		}
 		return new editorWorker();
 	},
 };
